+++ src/gen-bridge/PGenieGen.hs
@@ -1,0 +1,15 @@
+-- | Integration with generator adapters.
+module PGenieGen
+  ( Input,
+    Output,
+    Gen,
+    Location (..),
+    load,
+    bundle,
+  )
+where
+
+import PGenieGen.Bundle
+import PGenieGen.Load
+import PGenieGen.Location
+import PGenieGen.Model