+++ src/gen-bridge/PGenieGen/Dhall/Orphans.hs
@@ -1,0 +1,6 @@
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module PGenieGen.Dhall.Orphans where
+
+import PGenieGen.Dhall.Orphans.NonEmpty ()
+import PGenieGen.Dhall.Orphans.Path ()