+++ src/gen-bridge/PGenieGen/Model/Output.hs
@@ -1,0 +1,43 @@
+module PGenieGen.Model.Output
+  ( Output (..),
+    OutputOk (..),
+    Report (..),
+    File (..),
+  )
+where
+
+import Dhall qualified
+import PGenieGen.Dhall.Deriving qualified as Dhall.Deriving
+import PGenieGen.Dhall.Orphans ()
+import PGenieGen.Prelude
+
+-- | Diagnostic report with a context path and a message.
+data Report = Report
+  { path :: [Text],
+    message :: Text
+  }
+  deriving stock (Generic, Show, Eq)
+  deriving anyclass (Dhall.FromDhall, Dhall.ToDhall)
+
+-- | Successful generator output.
+data OutputOk = OutputOk
+  { warnings :: [Report],
+    value :: [File]
+  }
+  deriving stock (Generic, Show, Eq)
+  deriving anyclass (Dhall.FromDhall, Dhall.ToDhall)
+
+data Output
+  = OkOutput OutputOk
+  | ErrOutput Report
+  deriving stock (Generic, Show, Eq)
+  deriving
+    (Dhall.FromDhall, Dhall.ToDhall)
+    via (Dhall.Deriving.Codec (Dhall.Deriving.SumModifier "Output") Output)
+
+data File = File
+  { path :: Text,
+    content :: Text
+  }
+  deriving stock (Generic, Show, Eq)
+  deriving anyclass (Dhall.FromDhall, Dhall.ToDhall)