+++ src/gen-bridge/PGenieGen/Prelude.hs
@@ -1,0 +1,14 @@
+module PGenieGen.Prelude
+  ( module Exports,
+    Text,
+    Natural,
+    NonEmpty (..),
+    Generic,
+  )
+where
+
+import Data.List.NonEmpty (NonEmpty (..))
+import Data.Text (Text)
+import GHC.Generics (Generic)
+import Numeric.Natural (Natural)
+import Prelude as Exports